### 2023-12-30 #151 @Stew-McD
WasteAndMaterialFootprint - presubmission enquiry
Submitting Author: Stewart Charles McDowall (@Stew-McD)
Package Name:  WasteAndMaterialFootprint
One-Line Description of Package: A program for Life Cycle Assessment (LCA) calculations of supply chain waste and material footprints.
Repository Link (if existing):   https://github.com/Stew-McD/WasteAndMaterialFootprint

---

The WasteAndMaterialFootprint tool is a python package that allows one to easily and flexibly calculate the waste and material footprint of any product or service inside of life cycle assessment (LCA) databases. Currently is has been tested with ecoinvent 3.8, 3.9, 3.9.1 and 3.10. This is useful for sustainability assessments to calculate and consider the holistic impacts of waste and material demand for a given activity.

We partner with communities to support peer review with an additional layer of
checks that satisfy community requirements. If your package fits into an

This is a domain package to add value and utility for LCA practitioners and researchers in sustainability assessment. It adds some extra methods and features to the existing databases

No. This is built on the brightway2 LCA framework, and adds some extra stuff

[Conduct]: https://www.pyopensci.org/governance/CODE_OF_CONDUCT
