### 2024-01-04 #151 @NickleDave
Hi @Stew-McD! Welcome to pyOpenSci, and thank you for raising this presubmission issue.

The short answer is that, yes, WasteAndMaterialFootprint is in scope for pyOpenSci.

That said, you will need to meet the following criteria before you submit. These are taken from the [checklist](https://www.pyopensci.org/software-peer-review/how-to/editor-in-chief-guide.html#editor-checklist-template) we apply to all packages upon submission.
- [ ] **Documentation** The package has **sufficient online documentation to allow us to evaluate package function and scope *without installing the package***. This includes:

Re: documentation, you should immediately explain in both your README and docs *what* specifically the package does, and then show *how* you do it with the package. I'm having a hard time understanding what specifically the package adds on top of the existing brightway2 LCA framework. It might be because I'm not a domain expert, but it's not at all clear to me, even after spending significant time reading through your code and examples. The *first* thing someone will want to know is the *specific* methods and features that your package adds on top of what is in brightway2. Show them why they should care. I also can't tell *why* I would want to use the package. You should add narrative around usage so that a broader audience understands what the package makes possible--you have the start of this with your example but you will need a ot more to meet the criterion of "short tutorials" above.

Re: tests, you should use a testing framework (pytest). Our tutorials on tests are still in development, but in the meantime you may find this useful: https://learn.scientific-python.org/development/tutorials/test/

I would also strongly suggest you make the following changes, although they are not required:
- [ ] I would not use CapWords for module names (e.g. ExplodeDatabase.py) or functions (ExplodeDatabase) because this will confuse people. PEP8 conventions specify that [CapWords should be used for class names only.](https://peps.python.org/pep-0008/#package-and-module-names), and modules and functions should be all-lowercase with underscores (https://peps.python.org/pep-0008/#function-and-variable-names) except for maintaining backwards compatibility.
- [ ] I would give the package a shorter name. Why not name it `wmf` or `wamfoot` instead of asking a user to type the whole name and import it as an abbreviated form?
- [ ] The package has a huge number of dependencies. Is there a way that any of these can be made optional? Some of them seem like they are definitely not required (e.g. `cowsay`).
- [ ] I would remove the ["egg-info"](https://github.com/Stew-McD/WasteAndMaterialFootprint/tree/main/src/WasteAndMaterialFootprint.egg-info) from the git repository--I'm guessing this got added accidentally from a local development install? Usually I add *egg-info to the .gitignore to avoid doing this

Please let me know if all that is clear.

Please also know I don't mean to discourage you. I can tell that you are working hard on developing the package, and you are doing a lot of things right. But the core use of the package is not clear, and you're not following some key conventions. I want to help you fix that, and I want to make clear what our expectations are.
