### 2024-01-17 #151 @lwasser
hey there @Stew-McD i'm just following up on @NickleDave response above! if you have any questions on the items that david mentioned, you can either ask here or we also invite you to ask questions in our [discourse here](https://pyopensci.discourse.group/). The community can help you if you are interested in getting your package into shape for review! :) just let us know.
