### 2024-01-20 #151 @Stew-McD
Hi,
Thanks for the feedback! This is the first official package I've made so I am learning a lot as I go.
I'll work on some improvements based on your comments in the next couple of weeks.

All the best,
Stewart
