### 2024-07-09 #151 @Stew-McD
Yes! Thanks for following up here! Indeed, a name change as you suggested. Much more memorable now, I think. 🦖

We've just got the review back from the journal, so currently making some refinements.

I'll submit a new request once that is sorted.
:)
